# docs/openapi/ferrum.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ferrum",
    "description": "Implementation map of the Ferrum gateway (utoipa dump). Source of truth for GA4GH DRS/WES/TES/TRS/Beacon/htsget is each standard’s published OpenAPI — see docs/GA4GH.md. This file records what this process exposes, including Ferrum-only paths. Not a replacement spec. Not official GA4GH certification. Auth is a deployment choice. Behavioral proof against the published specs is HelixTest.",
    "contact": {
      "name": "Synaptic Four",
      "email": "contact@synapticfour.com"
    },
    "license": {
      "name": "BUSL-1.1",
      "url": "https://github.com/SynapticFour/Ferrum/blob/main/LICENSE"
    },
    "version": "0.3.1"
  },
  "servers": [
    {
      "url": "/",
      "description": "Paths are absolute from the Ferrum gateway origin (default http://127.0.0.1:8080)."
    }
  ],
  "paths": {
...
    "/ga4gh/tes/v1/tasks": {
      "get": {
        "tags": [
          "handlers"
        ],
        "operationId": "list_tasks",
        "parameters": [
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true
            }
          },
          {
            "name": "page_token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "handlers"
        ],
        "operationId": "create_task",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTaskResponse"
                }
              }
            }
          }
        }
      }
    },
    "/ga4gh/tes/v1/tasks/{id}": {
      "get": {
        "tags": [
          "handlers"
        ],
        "operationId": "get_task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      }
    },
    "/ga4gh/tes/v1/tasks/{id}:cancel": {
      "post": {
        "tags": [
          "handlers"
        ],
        "operationId": "cancel_task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTaskResponse"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      }
    },
...
    "/ga4gh/wes/v1/runs/{run_id}/tasks": {
      "get": {
        "tags": [
          "handlers"
        ],
        "summary": "GET /runs/{run_id}/tasks (paginated task logs)",
        "operationId": "list_tasks",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTaskRequest": {
        "type": "object",
        "description": "Create task request (POST /tasks body).",
        "required": [
          "executors"
        ],
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "executors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TesExecutor"
            }
          },
          "inputs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TesInput"
            },
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "outputs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TesOutput"
            },
            "nullable": true
          },
          "resources": {
            "nullable": true
          },
          "tags": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true
          },
          "volumes": {
            "type": "array",
            "items": {},
            "nullable": true
          }
        }
      },
      "CreateTaskResponse": {
        "type": "object",
        "description": "Create task response (task id).",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string"
          }
        }
      },
...
      "Task": {
        "type": "object",
        "description": "Full task (response).",
        "required": [
          "id",
          "state"
        ],
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "executors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TesExecutor"
            },
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "inputs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TesInput"
            },
            "nullable": true
          },
          "logs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskLog"
            },
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "outputs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TesOutput"
            },
            "nullable": true
          },
          "resources": {
            "nullable": true
          },
          "state": {
            "type": "string"
          },
          "tags": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true
          },
          "volumes": {
            "type": "array",
            "items": {},
            "nullable": true
          }
        }
      },
      "TaskListResponse": {
        "type": "object",
        "required": [
          "task_logs"
        ],
        "properties": {
          "next_page_token": {
            "type": "string",
            "nullable": true
          },
          "task_logs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskLog"
            }
          }
        }
      },
      "TaskLog": {
        "type": "object",
        "description": "Task log (with id).",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "cmd": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "end_time": {
            "type": "string",
            "nullable": true
          },
          "exit_code": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "start_time": {
            "type": "string",
            "nullable": true
          },
          "stderr": {
            "type": "string",
            "nullable": true
          },
          "stdout": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "TesExecutor": {
        "type": "object",
        "description": "Executor: image + command (TES 1.1).\n\n**Docker / Podman:** If the image defines an `ENTRYPOINT`, your `command` is passed as\nadditional arguments to that entrypoint. To run a shell wrapper (e.g. `bash -lc '…'`) instead,\nset [`Self::entrypoint`] explicitly (e.g. `[\"/bin/bash\", \"-lc\"]`) and put the script line in\n`command`, or clear the image entrypoint via `[\"/bin/sh\", \"-c\"]` plus a single joined command\nstring — see **`docs/TES-DOCKER-BACKEND.md`** in the Ferrum repository.",
        "required": [
          "image",
          "command"
        ],
        "properties": {
          "command": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "entrypoint": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Overrides the container image entrypoint (Docker `Entrypoint`). When set, `command` follows\nas argv after the entrypoint vector (Docker API semantics).",
            "nullable": true
          },
          "env": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true
          },
          "image": {
            "type": "string"
          },
          "stderr": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "stdin": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "stdout": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "workdir": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "TesInput": {
        "type": "object",
        "description": "Input file (url -> path in container).",
        "required": [
          "url",
          "path"
        ],
        "properties": {
          "path": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        }
      },
      "TesOutput": {
        "type": "object",
        "description": "Output path to capture (TES 1.1: `url` is the downloadable location).",
        "required": [
          "path"
        ],
        "properties": {
          "path": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  }
}